refactor: move CrCompareDependency to meta package Ref: #494
