+++ src/test/java/com/artipie/rpm/meta/CrCompareDependencyTest.java
@@ -2,10 +2,12 @@
  * The MIT License (MIT) Copyright (c) 2020-2021 artipie.com
  * https://github.com/artipie/rpm-adapter/LICENSE.txt
  */
-package com.artipie.rpm;
+package com.artipie.rpm.meta;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
+
+import com.artipie.rpm.meta.CrCompareDependency;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Assertions;
